requirements.txt
ccxt
python-dotenv
pandas